[lexical-playground] Refactor: Migrate FigmaPlugin to FigmaExtension
Converts the legacy React FigmaPlugin into a declarative FigmaExtension
using defineExtension, registering the FigmaNode and INSERT_FIGMA_COMMAND
through the extension system.
