+++ packages/lexical-playground/src/plugins/FigmaExtension/index.ts
@@ -6,12 +6,13 @@
  *
  */
 
-import type {JSX} from 'react';
-
-import {useLexicalComposerContext} from '@lexical/react/LexicalComposerContext';
 import {$insertNodeToNearestRoot} from '@lexical/utils';
-import {COMMAND_PRIORITY_EDITOR, createCommand, LexicalCommand} from 'lexical';
-import {useEffect} from 'react';
+import {
+  COMMAND_PRIORITY_EDITOR,
+  createCommand,
+  defineExtension,
+  LexicalCommand,
+} from 'lexical';
 
 import {$createFigmaNode, FigmaNode} from '../../nodes/FigmaNode';
 
@@ -19,15 +20,11 @@ export const INSERT_FIGMA_COMMAND: LexicalCommand<string> = createCommand(
   'INSERT_FIGMA_COMMAND',
 );
 
-export default function FigmaPlugin(): JSX.Element | null {
-  const [editor] = useLexicalComposerContext();
-
-  useEffect(() => {
-    if (!editor.hasNodes([FigmaNode])) {
-      throw new Error('FigmaPlugin: FigmaNode not registered on editor');
-    }
-
-    return editor.registerCommand<string>(
+export const FigmaExtension = defineExtension({
+  name: '@lexical/playground/Figma',
+  nodes: [FigmaNode],
+  register: (editor) =>
+    editor.registerCommand<string>(
       INSERT_FIGMA_COMMAND,
       (payload) => {
         const figmaNode = $createFigmaNode(payload);
@@ -35,8 +32,5 @@ export default function FigmaPlugin(): JSX.Element | null {
         return true;
       },
       COMMAND_PRIORITY_EDITOR,
-    );
-  }, [editor]);
-
-  return null;
-}
+    ),
+});
